+++ package.json
@@ -55,7 +55,7 @@
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
-    "d3": "^6.5.0",
+    "d3": "^5.16.0",
     "i18next": "^19.9.1",
     "moment": "^2.29.1",
     "polished": "^4.1.1",
